[src] Don't compare original columns with original one
Columns "cases_percentage" and "deaths_percentage",
in the report, are rounded to 1 decimal digit in such
a way that the sum of the values is exactly 100.0.
This means that comparing the full-precision versions
using np.allclose occasionally doesn't work.

Of course, I could implement that rounding myself instead
of using np.allclose. I don't think that's worth it.
In alternative, I could just use the original data without
recomputing the columns at all. Well, unfortunately, that
would require attention as well, since ISS guys have
forgotten decimal commas (dots) in the past, including
in the last report. Fixing a missing comma is not a huge
problem either; still, I prefer not trusting them and
recompute the columns.

Thus, for now, I'm going to:
- just stop enforcing np.allclose between recomputed and
  original percentage columns
- as an alternative sanity check, just ensure that the
  number in cases_percentage and deaths_percentage add up
  to 100.0.